Public API created
